requirements.txt
django>=2.1.2
djangorestframework
django-rest-swagger
coreapi
